refactor: move isCancelSort, createSortEvent function to /query/sort
